requirements.txt
fastapi
uvicorn
jinja2
tensorflow
pillow
python-multipart
requests
python-dotenv
authlib
starlette
